+++ src/ru/avalon/java/j20/labs/core/NumberBuilder.java
@@ -1,62 +1,31 @@
 package ru.avalon.java.j20.labs.core;
 
-import java.util.regex.Pattern;
+/**
+ * Абстрактное представление о строителе числа
+ */
+public interface NumberBuilder<E extends Number> {
 
-public class NumberBuilder {
-    private Double result = 0d;
-    private Double exponent = 0d;
-    private boolean hasSeparator = false;
-    private boolean empty = true;
-    private Pattern digitPattern = Pattern.compile("[0-9]");
-    private Pattern separatorPattern = Pattern.compile("[//.,//,]");
+    /**
+     * Добавляет элемент в строитель числа
+     */
+    void append(String digit);
 
+    /**
+     * Очищает строитель от всех ранее добавленных элементов
+     */
+    void delete();
 
-    public void append(String text) {
-        if (isDigit(text)) {
-            addDigit(text);
-            empty = false;
-        } else if (isSeparator(text) && !hasSeparator) {
-            exponent = -1d;
-            hasSeparator = true;
-            empty = false;
-        }
-    }
+    /**
+     * Возвращает результат построения числа
+     *
+     * @return число
+     */
+    E getResult();
 
-    private void addDigit(String text) {
-        if (!hasSeparator) {
-            result = result * Math.pow(10d, exponent) + Double.valueOf(text);
-            exponent = 1d;
-        } else {
-            result += Math.pow(10d, exponent) * Double.valueOf(text);
-            exponent--;
-        }
-    }
-
-    private boolean isDigit(String text) {
-        return digitPattern.matcher(text).find();
-    }
-
-    private boolean isSeparator(String text) {
-        return separatorPattern.matcher(text).find();
-    }
-
-    public Double getResult() {
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return result.toString();
-    }
-
-    public void delete() {
-        result = 0d;
-        exponent = 0d;
-        empty = true;
-        hasSeparator = false;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
+    /**
+     * Возвращает true если в строитель не добавлен ни один элемент
+     *
+     * @return
+     */
+    boolean isEmpty();
 }
